3x3 patterns: Include atari information for direct neighbors
This was not considered in the original Mogo paper, but seems to be a common practice nowadays and promises a good strength increase.

+++ board.h
@@ -67,7 +67,7 @@ typedef uint64_t hash_t;
 
 /* XXX: This really belongs in pattern3.h, unfortunately that would mean
  * a dependency hell. */
-typedef uint16_t hash3_t; // 3x3 pattern hash
+typedef uint32_t hash3_t; // 3x3 pattern hash
 
 
 /* Note that "group" is only chain of stones that is solidly
